requirements.txt
discord.py>=2.4.0
groq>=0.9.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
aiohttp>=3.9.0
httpx>=0.27.0
ormsgpack>=1.5.0
gtts>=2.5.0
python-pptx>=0.6.21
python-docx>=1.1.0
imageio-ffmpeg>=0.5.1
Pillow>=10.4.0
numpy>=1.26.0
opencv-python-headless>=4.10.0.84
requests>=2.32.0
pdfplumber>=0.11.0
PyPDF2>=3.0.1
google-api-python-client>=2.170.0
google-auth>=2.39.0
